add sensitive attribute symbol to attribute and add data accessor
